finetuning-requirements.txt
huggingface_hub==0.25.2
transformers
ipykernel
llama-cpp-python